moved Homepage.js
